refactor(foldkit): move the DevTools integration out of the boot closure
`startWith` resolved the `devTools` config, kept the view state and its
PubSub for Mounts, tracked Mount starts and ends, held the DevTools
store behind a `null` slot with the pause check, created the store with
its replay bridge, mounted the overlay, started the WebSocket bridge,
and recorded init, every Message, and every frame's Mounts, all inline
between the render and update code that uses them.

`makeDevToolsIntegration` in `devToolsIntegration.ts` now owns all of
that except the replay bridge. It takes the `devTools` config, `update`
and `maybeFreezeModel` for replay, and the queue's
`enqueueMessageEffect` for the WebSocket bridge, and returns the Mount
tracker and runtime, `drainMountEvents`, `readViewState` and
`setViewState`, `isPausedNow`, `installDevToolsStore`, `resumeDevTools`,
and the three recording helpers `recordInit`, `recordMessage`, and
`attachRenderedMounts`. Each recording helper carries the store null
check that used to sit at its call site, so the boot closure, the
update step, and the frame body call them unconditionally.

The store is created in a second step, `installDevToolsStore`, because
its replay bridge renders through `render`, reads and writes the render
frame's Model bookkeeping, and skips the pending View Transition, all
of which belong to the render side. That bridge stays in the boot
closure as the `renderReplay` and `markRenderPending` it hands to
`installDevToolsStore`, where the store creation used to be, and the
unit does nothing there when DevTools is off. Inside the bridge, the
rollback's `devToolsStore.resume` behind a null check became
`resumeDevTools`, which carries the same check.

The four `currentViewState` reads on the frame path became
`readViewState()` calls. `toCommandRecord` and the three `resolve*`
calls for the exclusion tags, the entry cap, and the keyframe interval
moved into the unit, so those three are resolved at boot instead of
when `makeRuntime` runs; all three are pure functions of the config.
The view-state PubSub is still created right after the Model PubSub.
The moved code is unchanged apart from formatting, the parameter and
return types the helpers gained, the `if` over the resolved config
becoming an `Option.match` whose `onNone` is `Effect.void`, the
`Effect.gen` and `Effect.suspend` wrappers that make `recordInit` and
`resumeDevTools` Effects, and a dropped `AnyCommand` cast on the
Commands passed to `recordMessage`, which was never needed.

No user-facing effect, so the changeset is empty.
